requirementes.txt
airtest==1.1.4
asn1crypto==0.24.0
attrs==19.1.0
Automat==0.7.0
bcrypt==3.1.6
cached-property==1.5.1
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
constantly==15.1.0
cryptography==2.7
cssselect==1.0.3
Deprecated==1.2.5
docker==3.7.2
docker-compose==1.24.0
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
facebook-wda==0.2.3
gitdb2==2.0.5
GitPython==2.1.11
hrpc==1.0.8
hyperlink==19.0.0
idna==2.7
incremental==17.5.0
Jinja2==2.10.1
jsonschema==2.6.0
lxml==4.3.4
MarkupSafe==1.1.1
mss==4.0.3
nose==1.3.7
numpy==1.16.4
opencv-contrib-python==3.4.2.17
paramiko==2.5.0
parsel==1.5.1
Pillow==6.0.0
pocoui==1.0.76
pyaml==19.4.1
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycparser==2.19
PyDispatcher==2.0.5
PyGithub==1.43.7
PyHamcrest==1.9.0
PyJWT==1.7.1
PyNaCl==1.3.0
pyOpenSSL==19.0.0
python-dateutil==2.8.0
python-gitlab==1.8.0
python-xlib==0.25
pywinauto==0.6.3
PyYAML==5.1.1
queuelib==1.5.0
requests==2.20.1
Scrapy==1.6.0
service-identity==18.1.0
six==1.12.0
smmap2==2.0.5
svn==0.3.46
texttable==0.9.1
Twisted==19.2.1
urllib3==1.24.3
w3lib==1.20.0
websocket-client==0.56.0
wrapt==1.11.1
zope.interface==4.6.0
